- Reaction roles rewritten completely. They now support multiple exclusivity groups per message and level requirements. However they can only be added one by one - Bot now support much higher XP values for global and server levels

+++ src/NadekoBot/Modules/Xp/_Common/LevelStats.cs
@@ -5,12 +5,12 @@ namespace NadekoBot.Modules.Xp;
 
 public class LevelStats
 {
-    public int Level { get; }
-    public int LevelXp { get; }
-    public int RequiredXp { get; }
-    public int TotalXp { get; }
+    public long Level { get; }
+    public long LevelXp { get; }
+    public long RequiredXp { get; }
+    public long TotalXp { get; }
 
-    public LevelStats(int xp)
+    public LevelStats(long xp)
     {
         if (xp < 0)
             xp = 0;
